feat: text editor dialog
